% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classAddress.R
\name{adrs}
\alias{adrs}
\title{A S3 class to represent the address for geocoding}
\usage{
adrs(street, housenumber, state, zip)
}
\arguments{
\item{street}{A string}

\item{housenumber}{A string}

\item{state}{A string}

\item{zip}{A string}
}
\description{
The S3 class address (\code{adrs}) is used to store the street, housenumber,
zip code and the state of an \code{adrs} object.
}

