% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodFormatting.R
\name{formatting}
\alias{formatting}
\title{Formatting of address or coordinates.}
\usage{
formatting(x)
}
\arguments{
\item{x}{Object of class \code{adrs} or \code{coord}.}
}
\description{
Returns a formatted string of an \code{adrs} or \code{coord} object.
}
\examples{
x <- adrs('Linköping University, Linköping')
formatting(x)

y <- coord(50.231, -17.8239)
formatting(y)

}

