% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classGoogleGeo.R
\name{geocode.coord}
\alias{geocode.coord}
\title{Reverse geocoding of coordinates.}
\usage{
\method{geocode}{coord}(coordinates)
}
\arguments{
\item{x}{Object of class \code{\link{coord}}.}
}
\description{
This function returns the result of reverse geocoding. Reverse geocoding
describes the process of finding the address which belongs to given
coordinates.
}
\examples{
x <- adrs('Linköping Universitetet, Linköping')
geocode(x)
y <- coord(58.40208, 15.57901)
geocode(y)

}

