% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classGoogleGeo.R
\name{geocode.googlegeo_api}
\alias{geocode.googlegeo_api}
\title{(Reverse) Geocoding of an address (coordinates).}
\usage{
\method{geocode}{googlegeo_api}(x)
}
\arguments{
\item{x}{Object of class \code{googlegeo_api}.}
}
\description{
This function returns the result of (reverse) geocoding. (Reverse) Geocoding
describes the process of finding the coordinates to a given address
(the address which belongs to given coordinates).
}
\examples{
x <- adrs('Linköping Universitetet, Linköping')
geocode(googlegeo_api(x))
y <- coord(58.40208, 15.57901)
geocode(googlegeo_api(y))
}

