% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classGoogleGeo.R
\name{geocode}
\alias{geocode}
\title{(Reverse) Geocoding of an address (coordinates).}
\usage{
geocode(x)
}
\arguments{
\item{x}{Object of class \code{googlegeo_api}.}
}
\description{
This function returns the result of (reverse) geocoding. (Reverse) Geocoding
describes the process of finding the coordinates to a given address
(the address which belongs to given coordinates).
}
\examples{
x <- adrs('Linköping Universitetet, Linköping')
y <- googlegeo_api(x)
geocode(y)

}

