% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classGoogleGeo.R
\name{googlegeo_api}
\alias{googlegeo_api}
\title{Runs Google Geo API.}
\usage{
googlegeo_api(x)
}
\arguments{
\item{x}{a \code{\link{adrs}} or \code{\link{coord}} object}
}
\value{
\code{googlegeo_api} object that contains result returned by Google
Geo API.
}
\description{
This function runs Runs Google Geo API on \code{\link{adrs}} or
\code{\link{coord}} object.
}

