% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoogleGeoAPI.R
\docType{package}
\name{GoogleGeoAPI}
\alias{GoogleGeoAPI}
\alias{GoogleGeoAPI-package}
\title{GoogleGeoAPI: A package for laboratory work 4 of the Advanced R Programming
course.}
\description{
The package provides functions for (reverse) geocoding. (Reverse) geocoding
will give you the coordinates for a given address (the addres for given
coordinates).
}
\details{
{\code{LinregClass}}:\itemize{
\item{\code{\link{adrs}} returns an S3 class object of type \code{adrs}.}
\item{\code{\link{coord}} returns an S3 class object of type \code{coord}.}
\item{\code{\link{googlegeo_api}} returns an S3 class object of type \code{googlegeo_api}.}
\item{\code{\link{geocode}} returns the result of (reverse) geocoding.}
}
}

