% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classAddress.R
\name{adrs}
\alias{adrs}
\title{A S3 class to represent the address for a geocoding request.}
\usage{
adrs(a)
}
\arguments{
\item{x}{A string}
}
\description{
Creates an object of the class \code{adrs}. The \code{adrs} object contains a
single string representing the address of a place.
}
\examples{
x <- adrs('Linköpings Universitet, 58183 Linköping')
print(x)

}
\seealso{
\href{https://developers.google.com/maps/documentation/#'geocoding/
intro?hl=de}{Google Geo API Description}
}

