% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classAddress.R
\name{adrs}
\alias{adrs}
\title{Lorem ipsulum}
\usage{
adrs(street, housenumber, state)
}
\arguments{
\item{street}{A string}

\item{housenumber}{A string}

\item{state}{A string}
}
\description{
Lorem ipsulum
}

