% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classGoogleGeo.R
\name{googlegeo_api}
\alias{googlegeo_api}
\title{Runs Google Geo API on \code{\link{adrs}} or \code{\link{coord}} object.}
\usage{
googlegeo_api(x)
}
\arguments{
\item{x}{a \code{\link{adrs}} or \code{\link{coord}} object}
}
\value{
JSON object returned by Google Geo API.
}
\description{
Runs Google Geo API on \code{\link{adrs}} or \code{\link{coord}} object.
}

