% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classGoogleGeo.R
\name{googlegeo_api}
\alias{googlegeo_api}
\title{Run Google Geo API on \code{\link{adrs}} or \link{coord} object.}
\usage{
googlegeo_api(x)
}
\arguments{
\item{x}{asdfasdf}
}
\description{
Run Google Geo API on \code{\link{adrs}} or \link{coord} object.
}

